Use chrome instead of weasyprint
This major update changes how PDFs are built. Instead of requiring
weasyprint we use Google Chrome (or Chromium). Weasyprint turns out to
be non-trivial to install (especially on Windows and Apple Silicon).

This change is made by adding functionality to launch Chrome in headless
mode to resume.py. This requires lots of special cases and options to
get things to work on all the platforms I've tested.

resume.py is now a "real" command line program with --no-html,
--no-pdf, --quiet and --chrome-path options.

The script uses Chromium if it is available, falling back to the evil version if
Chromium is not found. It searches likely paths for the Chromium/Chrome
executable, but you can specify the path with the --chrome-path option.

This commit also adds a GitHub action to build HTML and PDF on Linux,
Windows and macOS. Inspection of the build artifacts is useful for
testing. It is related to #1, but is probably not ready for use by
regular users.

Closes #4, #6, #8