fix: use absolute path to fix pdf output on windows
Based on user reports (see #14), Chrome PDF output to fail in some
circumstances on Windows. It doesn't fail in the CI builds, so we don't
have a true root cause. :shrug:

The fix (or at least benign change!) is to pass Chrome an absolute path
to the PDF.

Credit to #18 for this fix! I've chosen to cherry-pick that fix (that PR
also addresses a separate issue).