feat: add --no-sandbox parameter to Chromium
When running resume.md with GitLab CI, the process is running inside a container, and hence Chromium require a specific configuration in order to run as a headless browser. The process is also often running as root and this has a specific issue:

```
Wrote /lab/resume.html
Found Chrome or Chromium at /usr/bin/google-chrome
[0715/191521.895591:ERROR:zygote_host_impl_linux.cc(90)] Running as root without --no-sandbox is not supported. See https://crbug.com/638180.
```

It is hence a common workaround to run Chromium with `--no-sandbox`.